[updater] refectored as requested by reviewer
