+++ bundles/org.openhab.core.io.rest.update/.classpath
