+++ bundles/org.openhab.core.io.rest.update/src/main/java/org/openhab/core/io/rest/update/internal/dto/UpdaterStatusDTO.java
@@ -1,0 +1,26 @@
+/**
+ * Copyright (c) 2010-2021 Contributors to the openHAB project
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.openhab.core.io.rest.update.internal.dto;
+
+/**
+ * This is a DTO for OpenHAB self updating features.
+ *
+ * @author Andrew Fiddian-Green - Initial contribution
+ */
+public class UpdaterStatusDTO {
+    public OpenHabVersionDTO actualVersion;
+    public Integer latestVersionCount;
+    public OpenHabVersionDTO[] latestVersions;
+    public String targetNewVersionType;
+    public String newVersionAvailable;
+}
