+++ bundles/org.openhab.core.io.rest.update/src/main/resources/scripts/MacUpdater.txt
