+++ bundles/org.openhab.core.io.rest.update/src/main/resources/scripts/WindowsUpdater.txt
